adding transfer test
- Should transfer balance from sender to receiver
- Should return true on a 0 value transfer
vanductai/coincapo-pm#94

+++ contracts/Token/CAP.sol
@@ -5,7 +5,7 @@ import "./UnlimitedAllowanceToken.sol";
 contract CAP is UnlimitedAllowanceToken {
     function CAP() {
     	decimals = 18;
-    	totalSupply = 10**27; // 1 billion tokens, 18 decimal places
+    	totalSupply = uint(10**27); // 1 billion tokens, 18 decimal places
     	name = "CAPO Dex Token";
     	symbol = "CAP";
     	balances[msg.sender] = totalSupply;
